Example of refactoring service components to inherit from a general ServiceComponent
