+++ src/components/service/DefaultService.js
@@ -1,89 +1,19 @@
 import React from 'react';
 import {hasOwnDefinedProperty} from '../../util'
+import ServiceComponent from './ServiceComponent'
 
-export default class DefaultService extends React.Component {
+export default class DefaultService extends ServiceComponent {
 
     constructor(props) {
         super(props);
-        this.submitAction = this.submitAction.bind(this);
-        this.handleServiceName = this.handleServiceName.bind(this);
-        this.handleFormUpdate = this.handleFormUpdate.bind(this);
-        this.handleResponse = this.handleResponse.bind(this);
-
-        this.state = {
-            serviceName: undefined,
-            methodName: undefined,
-            response: undefined,
-            paramString: "{}"
-        };
-        this.isComplete = false;
-        this.serviceMethods = [];
-        this.allServices = [];
-        this.methodsForAllServices = [];
-        this.parseProps(props);
-    }
-
-    parseProps(nextProps) {
-        this.isComplete = nextProps.isComplete;
-        this.parseServiceSpec(nextProps.serviceSpec);
-    }
-    componentWillReceiveProps(nextProps) {
-        if(this.isComplete !== nextProps.isComplete) {
-            this.parseProps(nextProps);
-        }
-    }
-    parseServiceSpec(serviceSpec) {
-        const packageName = Object.keys(serviceSpec.nested).find(key =>
-            typeof serviceSpec.nested[key] === "object" &&
-            hasOwnDefinedProperty(serviceSpec.nested[key], "nested"));
-
-        var objects = undefined;
-        var items = undefined;
-        if (typeof packageName !== 'undefined') {
-            items = serviceSpec.lookup(packageName);
-            objects = Object.keys(items);
-        } else {
-            items = serviceSpec.nested;
-            objects = Object.keys(serviceSpec.nested);
-        }
-
-        this.allServices.push("Select a service");
-        this.methodsForAllServices = [];
-        objects.map(rr => {
-            if (typeof items[rr] === 'object' && items[rr] !== null && items[rr].hasOwnProperty("methods")) {
-                this.allServices.push(rr);
-                this.methodsForAllServices.push(rr);
-
-                var methods = Object.keys(items[rr]["methods"]);
-                methods.unshift("Select a method");
-                this.methodsForAllServices[rr] = methods;
-            }
-        })
-    }
-
-    handleFormUpdate(event) {
-        this.setState({
-            [event.target.name]: event.target.value
-        });
-    }
-
-    handleServiceName(event) {
-        let strService = event.target.value;
-        this.setState({
-            serviceName: strService
-        });
-        this.serviceMethods.length = 0;
-        if (typeof strService !== 'undefined' && strService !== 'Select a service') {
-            let data = Object.values(this.methodsForAllServices[strService]);
-            if (typeof data !== 'undefined') {
-                this.serviceMethods= data;
-            }
-        }
+        var customState = {paramString: "{}"};
+        Object.assign(this.state, customState);
     }
 
     handleResponse(response)
     {
         this.setState({response: JSON.stringify(response)});
+        this.isComplete = true;
     }
 
     submitAction() {
@@ -144,22 +74,6 @@ export default class DefaultService extends React.Component {
                 <p style={{fontSize: "13px"}}>Response from service is {this.state.response} </p>
             </div>
         );
-    }
-
-    render() {
-        if (this.isComplete)
-            return (
-                <div>
-                    {this.renderComplete()}
-                </div>
-            );
-        else {
-            return (
-                <div>
-                    {this.renderForm()}
-                </div>
-            )
-        }
     }
 
 }
