+++ src/components/service/ServiceComponent.js
@@ -1,0 +1,102 @@
+import React from 'react';
+import {hasOwnDefinedProperty} from '../../util'
+
+export default class ServiceComponent extends React.Component {
+
+    constructor(props) {
+        super(props);
+        this.submitAction = this.submitAction.bind(this);
+        this.handleServiceName = this.handleServiceName.bind(this);
+        this.handleFormUpdate = this.handleFormUpdate.bind(this);
+        this.handleResponse = this.handleResponse.bind(this);
+
+        this.state = {
+            serviceName: undefined,
+            methodName: undefined,
+            response: undefined,
+        };
+        this.serviceMethods = [];
+        this.allServices = [];
+        this.methodsForAllServices = [];
+        this.parseServiceSpec(props.serviceSpec);
+        this.isComplete = false;
+    }
+
+    parseServiceSpec(serviceSpec) {
+        const packageName = Object.keys(serviceSpec.nested).find(key =>
+            typeof serviceSpec.nested[key] === "object" &&
+            hasOwnDefinedProperty(serviceSpec.nested[key], "nested"));
+
+        var objects = undefined;
+        var items = undefined;
+        if (typeof packageName !== 'undefined') {
+            items = serviceSpec.lookup(packageName);
+            objects = Object.keys(items);
+        } else {
+            items = serviceSpec.nested;
+            objects = Object.keys(serviceSpec.nested);
+        }
+
+        this.allServices.push("Select a service");
+        this.methodsForAllServices = [];
+        objects.map(rr => {
+            if (typeof items[rr] === 'object' && items[rr] !== null && items[rr].hasOwnProperty("methods")) {
+                this.allServices.push(rr);
+                this.methodsForAllServices.push(rr);
+
+                var methods = Object.keys(items[rr]["methods"]);
+                methods.unshift("Select a method");
+                this.methodsForAllServices[rr] = methods;
+            }
+        })
+    }
+
+    handleFormUpdate(event) {
+        this.setState({
+            [event.target.name]: event.target.value
+        });
+    }
+
+    handleServiceName(event) {
+        let strService = event.target.value;
+        this.setState({
+            serviceName: strService
+        });
+        this.serviceMethods.length = 0;
+        if (typeof strService !== 'undefined' && strService !== 'Select a service') {
+            let data = Object.values(this.methodsForAllServices[strService]);
+            if (typeof data !== 'undefined') {
+                this.serviceMethods= data;
+            }
+        }
+    }
+
+    renderForm() {
+        return (
+            <React.Fragment><div>Implement Me!</div></React.Fragment>
+        )
+    }
+
+    renderComplete() {
+        return (
+            <div>Implement Me!</div>
+        );
+    }
+
+    render() {
+        if (this.isComplete)
+            return (
+                <div>
+                    {this.renderComplete()}
+                </div>
+            );
+        else {
+            return (
+                <div>
+                    {this.renderForm()}
+                </div>
+            )
+        }
+    }
+
+}
